Added the "Compatibility" chapter to the READMEs
In order to make the support for different Elasticsearch versions more clear,
a special "Compatibility" chapter has been added to the main README and
individual gem READMEs.

The versioning scheme follows the scheme for the Elasticsearch Ruby client,
ie. the version number matches the supported major version of Elasticsearch.

See:

* https://github.com/elastic/elasticsearch-ruby/commit/3bdb385